Documentation: Generate local TOC on demand
All pages are generated with a local TOC, which is pointless on simple
pages such as the front page, or other pages containing little content.
Use the .. contents:: :local: directive instead to generate the local
TOC on demand, and remove the automatic local TOC generation. Only the
Docs page uses a local TOC.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>