libcamera: v4l2_videodevice: Add FrameBuffer interface
Add a new interface in parallel with the existing Buffer implementation
to also support FrameBuffer. The reason it's added in parallel is to aid
in the migration from Buffer to FrameBuffer throughout libcamera. With
this change discrete parts of libcamera can be migrated and tested
independently.

As the new interface is added in parallel there are some oddities in
this change which will be undone in a follow up patch once libcamera
have migrated away from the Buffer interface.

- There is a nasty hack in V4L2VideoDevice::bufferAvailable(). It is
  needed to allow both interfaces to exist and function at the same
  time. The idea is if buffers are allocated using the FrameBuffer
  interface V4L2VideoDevice::cache_ is set and we know to call the
  FrameBuffer 'buffer ready' signal, and likewise if it's not to call
  the Buffer variant.

- There is some code duplication between the two interfaces as they aim
  to solve the same thing in slightly different ways. As all Buffer
  related code is soon to be removed no effort to create code sharing
  between them have been made.

- Some function and variables which can't be distinguished by their
  argument types have been given a frameBuffer prefix instead of a
  buffer prefix. They are clearly documented in the code and will be
  renamed to the correct buffer prefix when the Buffer interface is
  removed.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>